+++ apps/web/src/app/mypage/edit/_components/EditForm.tsx
@@ -9,8 +9,8 @@ import Spacing from '@/components/spacing';
 import { ROUTES } from '@/consts/route';
 import { useGetMe } from '@/hooks/useGetMe';
 import { useNicknameValidation } from '@/hooks/useNicknameValidation';
+import { usePatchMe } from '@/hooks/usePatchMe';
 
-import { usePatchMe } from '../_hooks/usePatchMe';
 import NicknameField from './NicknameField';
 import ProfileImageField from './ProfileImageField';
